completed

+++ src/main/java/com/example/admin/service/PostServiceImpl.java
@@ -1,19 +1,25 @@
 package com.example.admin.service;
 
 import com.example.admin.entity.Post;
+import com.example.admin.entity.Thread;
 import com.example.admin.entity.User;
+import com.example.admin.logic.ThreadLogic;
+import com.example.admin.logic.UserLogic;
 import com.example.admin.repository.PostRepository;
+import com.example.admin.request.PostCreateRequest;
 import com.example.admin.response.PostResponse;
-import com.example.admin.response.UserResponse;
-import com.example.admin.utility.TimestampManager;
-import org.hibernate.boot.model.naming.IllegalIdentifierException;
+import com.example.admin.security.MyUserDetails;
+import com.example.admin.utility.SecurityUtil;
+import com.example.admin.utility.TimestampUtil;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.context.properties.source.InvalidConfigurationPropertyValueException;
+import org.springframework.security.access.AccessDeniedException;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import javax.servlet.http.HttpServletRequest;
 import java.util.ArrayList;
-import java.util.Date;
 import java.util.List;
 
 @Service
@@ -24,48 +30,81 @@ public class PostServiceImpl implements PostService{
     PostRepository postRepository;
 
     @Autowired
-    TimestampManager timestampManager;
+    TimestampUtil timestampManager;
+
+    @Autowired
+    SecurityUtil securityUtil;
+
+    @Autowired
+    UserLogic userLogic;
+
+    @Autowired
+    ThreadLogic threadLogic;
 
     @Override
-    public void create(String ip ,User user, String content) {
+    public void createPost(Authentication auth, HttpServletRequest req, PostCreateRequest reqBody , Long threadId) {
+
+        if(auth == null || req == null || reqBody == null){return;}
+        MyUserDetails userDetails = (MyUserDetails)auth.getPrincipal();
+
+        if(userDetails == null){return;}
         Post post = new Post();
-        post.setIp(ip);
+        User user = userLogic.getEntitiyByUserId(userDetails.getUserId());
         post.setUser(user);
-        post.setContent(content);
+        Thread thread = threadLogic.getEntityByThreadId(threadId);
+        post.setThread(thread);
+        post.setIp(req.getRemoteAddr());
+        post.setContent(reqBody.getContent());
+        post.setValid(true);
         post.setCreatedAt(timestampManager.getNow());
         postRepository.save(post);
+
     }
 
     @Override
     public List<PostResponse> getAllResponses() {
+
         List<Post> posts = postRepository.findAll();
         List<PostResponse> postResponses = new ArrayList<>();
         //投稿情報をレスポンス用のリストに格納する
-        for(Post post :posts){
+        posts.forEach((Post post)->{
             postResponses.add(this.getResponse(post));
-        }
+        });
+
         return postResponses;
     }
 
     @Override
-    public List<PostResponse> getAllResponsesByUserId(Integer userId) {
+    public List<PostResponse> getAllResponsesByUserId(Long userId) {
+
         List<Post> posts = postRepository.findAllByUserId(userId);
         List<PostResponse> postResponses = new ArrayList<>();
-        for(Post post : posts){
+        posts.forEach((Post post)->{
             postResponses.add(new PostResponse(post));
-        }
+        });
+
         return postResponses;
     }
 
     @Override
-    public PostResponse getResponseByPostId(Integer postId) {
+    public List<PostResponse> getAllResponseByThreadId(Long threadId) {
+        List<Post> posts = postRepository.findAllByThreadId(threadId);
+        List<PostResponse> postResponses = new ArrayList<>();
+        posts.forEach((Post post)->{
+            postResponses.add(new PostResponse(post));
+        });
+        return postResponses;
+    }
+
+    @Override
+    public PostResponse getResponseByPostId(Long postId) {
 
         Post post = this.getEntityByPostId(postId);
         return new PostResponse(post);
 
     }
 
-    public Post getEntityByPostId(Integer postId) {
+    public Post getEntityByPostId(Long postId) {
         Post post = postRepository.findById(postId)
                 .orElseThrow(()-> new IllegalArgumentException());
         return post;
@@ -76,30 +115,37 @@ public class PostServiceImpl implements PostService{
     }
 
     @Override
-    public void validate(Integer postId) {
+    public void validateByPostId(Long postId) {
         Post post = this.getEntityByPostId(postId);
-        post.setIsValid(true);
-        post.setUpdatedAt(timestampManager.getNow());
-        postRepository.save(post);
+        if(!post.isValid()) {
+            post.setValid(true);
+            post.setUpdatedAt(timestampManager.getNow());
+            postRepository.save(post);
+        }
     }
 
     @Override
-    public void invalidate(Integer postId) {
+    public void invalidateByPostId(Long postId) {
         Post post = this.getEntityByPostId(postId);
-        post.setIsValid(false);
-        post.setUpdatedAt(timestampManager.getNow());
-        postRepository.save(post);
+        if(post.isValid()) {
+            post.setValid(false);
+            post.setUpdatedAt(timestampManager.getNow());
+            postRepository.save(post);
+        }
     }
 
     @Override
-    public void delete(Integer postId) {
+    public void deleteByPostId(Long postId , Authentication auth) {
+        if(auth == null){return;}
+        MyUserDetails myUserDetails = (MyUserDetails) auth.getPrincipal();
         Post post = this.getEntityByPostId(postId);
+        //認証を受けたユーザーが、Postを投稿したユーザーと一致せず、かつユーザーがADMINではない場合はアクセスを拒否
+        if(!securityUtil.isAuthIdEqualPathId(myUserDetails.getUserId() , post.getUser().getUserId())
+                && !securityUtil.isAdmin(myUserDetails.getAuthorities())
+        ){
+            throw new AccessDeniedException("");
+        }
         postRepository.delete(post);
-    }
-
-    @Override
-    public void deleteAllByUserId(Integer userId) {
-        postRepository.deleteAllByUserId(userId);
     }
 
 }
